Move resource factories into util packages.
