+++ collections/src/main/java/io/atomix/collections/DistributedMap.java
@@ -16,6 +16,7 @@
 package io.atomix.collections;
 
 import io.atomix.collections.state.MapCommands;
+import io.atomix.collections.util.DistributedMapFactory;
 import io.atomix.copycat.client.CopycatClient;
 import io.atomix.resource.AbstractResource;
 import io.atomix.resource.ReadConsistency;
