+++ collections/src/main/java/io/atomix/collections/DistributedSet.java
@@ -16,6 +16,7 @@
 package io.atomix.collections;
 
 import io.atomix.collections.state.SetCommands;
+import io.atomix.collections.util.DistributedSetFactory;
 import io.atomix.copycat.client.CopycatClient;
 import io.atomix.resource.AbstractResource;
 import io.atomix.resource.ReadConsistency;
