+++ collections/src/main/java/io/atomix/collections/util/DistributedQueueFactory.java
@@ -13,37 +13,38 @@
  * See the License for the specific language governing permissions and
  * limitations under the License
  */
-package io.atomix.group;
+package io.atomix.collections.util;
 
 import io.atomix.catalyst.serializer.SerializableTypeResolver;
+import io.atomix.collections.DistributedQueue;
+import io.atomix.collections.state.QueueCommands;
+import io.atomix.collections.state.QueueState;
 import io.atomix.copycat.client.CopycatClient;
-import io.atomix.group.state.GroupCommands;
-import io.atomix.group.state.GroupState;
 import io.atomix.resource.ResourceFactory;
 import io.atomix.resource.ResourceStateMachine;
 
 import java.util.Properties;
 
 /**
- * Distributed group factory.
+ * Distributed queue factory.
  *
  * @author <a href="http://github.com/kuujo>Jordan Halterman</a>
  */
-public class DistributedGroupFactory implements ResourceFactory<DistributedGroup> {
+public class DistributedQueueFactory implements ResourceFactory<DistributedQueue<?>> {
 
   @Override
   public SerializableTypeResolver createSerializableTypeResolver() {
-    return new GroupCommands.TypeResolver();
+    return new QueueCommands.TypeResolver();
   }
 
   @Override
   public ResourceStateMachine createStateMachine(Properties config) {
-    return new GroupState(config);
+    return new QueueState(config);
   }
 
   @Override
-  public DistributedGroup createInstance(CopycatClient client, Properties options) {
-    return new DistributedGroup(client, options);
+  public DistributedQueue<?> createInstance(CopycatClient client, Properties options) {
+    return new DistributedQueue<>(client, options);
   }
 
 }
