+++ concurrent/src/main/java/io/atomix/concurrent/DistributedLock.java
@@ -16,6 +16,7 @@
 package io.atomix.concurrent;
 
 import io.atomix.concurrent.state.LockCommands;
+import io.atomix.concurrent.util.DistributedLockFactory;
 import io.atomix.copycat.client.CopycatClient;
 import io.atomix.resource.AbstractResource;
 import io.atomix.resource.Resource;
