+++ variables/src/main/java/io/atomix/variables/DistributedLong.java
@@ -19,6 +19,7 @@ import io.atomix.copycat.client.CopycatClient;
 import io.atomix.resource.ReadConsistency;
 import io.atomix.resource.ResourceTypeInfo;
 import io.atomix.variables.state.LongCommands;
+import io.atomix.variables.util.DistributedLongFactory;
 
 import java.util.Properties;
 import java.util.concurrent.CompletableFuture;
